new changes

+++ lib/view/helpers/colors.dart
@@ -3,5 +3,5 @@ import 'package:flutter/material.dart';
 
 Color primeColor= Color(0xffe5283c);
 Color whiteColor=  Colors.white;
-int maxSize = 18;
-int minSize = 14;
+double maxSize = 18;
+double minSize = 14;
